code_comparity/requirements.txt
absl-py==2.0.0
accelerate==0.25.0
aiohttp==3.8.4
aioredis==2.0.1
aiosignal==1.3.1
altair==5.1.1
annotated-types==0.5.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.5.0
asttokens==2.4.0
astunparse==1.6.3
async-timeout==4.0.2
asyncio==3.4.3
attrs==23.1.0
Authlib==1.3.0
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.2
bitsandbytes==0.41.0
blinker==1.6.2
boilerpy3==1.0.7
boto3==1.26.150
botocore==1.29.150
cached-property==1.5.2
cachetools==5.3.1
cattrs==23.2.3
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.21
click==8.1.7
cloud-init==23.2.1
cmake==3.28.1
colorama==0.4.4
coloredlogs==15.0.1
comm==0.1.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
construct==2.5.3
contourpy==1.2.0
cryptography==3.4.8
ctransformers==0.2.27
cycler==0.12.1
dataclasses-json==0.5.7
datasets==2.14.0
dbus-python==1.2.18
debugpy==1.8.0
decorator==5.1.1
deepface==0.0.75
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.7.0
distro-info==1.1+ubuntu0.1
dlib==19.24.2
dnspython==2.4.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
einops==0.6.1
Events==0.5
exceptiongroup==1.1.3
executing==2.0.0
face-recognition==1.3.0
face-recognition-models==0.3.0
faiss-cpu==1.7.4
farm-haystack==1.23.0
fastapi==0.103.2
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.13.1
fire==0.5.0
Flask==3.0.0
Flask-Login==0.6.3
flatbuffers==23.5.26
fonttools==4.44.3
frozenlist==1.3.3
fsspec==2023.10.0
future==0.18.3
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.37
google-auth==2.23.4
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googleapis-common-protos==1.62.0
greenlet==3.0.2
grpcio==1.59.2
gTTS==2.4.0
h11==0.14.0
h5py==3.10.0
haystack==0.42
haystack-ai==2.0.0b3
httpcore==1.0.2
httplib2==0.20.2
httptools==0.6.1
httpx==0.25.2
huggingface-hub==0.19.4
humanfriendly==10.0
hyperlink==21.0.0
idna==3.4
imageio==2.31.5
imgkit==1.2.3
importlib-metadata==6.8.0
importlib-resources==6.1.1
imutils==0.5.4
incremental==21.3.0
inexactsearch==1.0.2
inflect==7.0.0
ipykernel==6.25.2
ipython==8.13.0
itsdangerous==2.1.2
jedi==0.19.1
jeepney==0.7.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.0
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
keras==2.13.1
keyring==23.5.0
kiwisolver==1.4.5
kubernetes==28.1.0
langchain==0.0.352
langchain-community==0.0.4
langchain-core==0.1.1
langchainplus-sdk==0.0.8
langsmith==0.0.72
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy-imports==0.3.1
lazy_loader==0.3
libclang==16.0.6
lit==17.0.6
llama-index==0.9.16.post1
llama_cpp_python==0.2.24
llvmlite==0.41.0
loguru==0.7.2
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.8
mmh3==4.0.1
monotonic==1.6
more-itertools==8.10.0
mpmath==1.3.0
mtcnn==0.1.1
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
mysql-connector==2.2.9
mysql-connector-python==8.2.0
nest-asyncio==1.5.8
netifaces==0.11.0
networkx==3.1
nltk==3.8.1
num2words==0.5.13
numba==0.58.0
numexpr==2.8.4
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
onnxruntime==1.16.0
openai==1.6.0
openai-whisper @ git+https://github.com/openai/whisper.git@e58f28804528831904c3b6f2c0e473f346223433
openapi-schema-pydantic==1.2.4
opencv-contrib-python==4.7.0.68
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
opentelemetry-api==1.22.0
opentelemetry-exporter-otlp-proto-common==1.22.0
opentelemetry-exporter-otlp-proto-grpc==1.22.0
opentelemetry-instrumentation==0.43b0
opentelemetry-instrumentation-asgi==0.43b0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-proto==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-semantic-conventions==0.43b0
opentelemetry-util-http==0.43b0
opt-einsum==3.3.0
outcome==1.3.0.post0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pefile==2023.2.7
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
pinecone-client==2.2.2
platformdirs==3.10.0
pooch==1.7.0
posthog==3.1.0
prompt-toolkit==3.0.39
prompthub-py==4.0.0
protobuf==4.24.3
psutil==5.9.5
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pyenchant==3.2.2
Pygments==2.16.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
PyMatting==1.1.8
Pympler==1.0.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pypdf==3.17.3
PyPDF2==3.0.1
PyPika==0.48.9
pyre-extensions==0.0.29
pyrsistent==0.19.3
pyserial==3.5
PySocks==1.7.1
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
python-multipart==0.0.6
python-ptrace==0.9.8
PythonTurtle==0.3.2
pytube==12.1.0
pytz==2023.3.post1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.1
quantulum3==0.9.0
rank-bm25==0.2.2
redis==3.5.3
referencing==0.30.2
regex==2023.10.3
rejson==0.5.6
rembg==2.0.53
replicate==0.22.0
requests==2.31.0
requests-cache==0.9.8
requests-oauthlib==1.3.1
retina-face==0.0.13
rich==13.6.0
rpds-py==0.10.3
rsa==4.9
s3transfer==0.6.1
safetensors==0.4.1
scikit-image==0.21.0
scikit-learn==1.0.2
scipy==1.10.1
SecretStorage==3.3.1
selenium==4.16.0
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
service-identity==18.1.0
silpa-common==0.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
sortedcontainers==2.4.0
sos==4.4
sounddevice==0.4.6
soundex==1.1.3
soupsieve==2.5
SpeechRecognition==3.10.0
speedtest-cli==2.1.3
spellchecker==0.4
SQLAlchemy==2.0.15
sseclient-py==1.8.0
ssh-import-id==5.11
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.1
streamlit-chat==0.0.2.2
streamlit-lottie==0.0.3
sympy==1.12
systemd-python==234
tenacity==8.2.3
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
texttable==1.6.4
threadpoolctl==3.2.0
tifffile==2023.7.10
tiktoken==0.5.1
tokenizers==0.15.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.10.1
transformers==4.35.2
trio==0.24.0
trio-websocket==0.11.1
triton==2.0.0
Twisted==22.1.0
typer==0.9.0
types-protobuf==4.24.0.4
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.0.1
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36.1
unattended-upgrades==0.1
url-normalize==1.4.3
urllib3==2.0.5
uvicorn==0.23.2
uvloop==0.19.0
validators==0.22.0
wadllib==1.3.6
watchdog==3.0.0
watchfiles==0.21.0
wcwidth==0.2.8
weaviate-client==3.26.0
websocket-client==1.2.3
websockets==12.0
Werkzeug==3.0.1
whisper==1.1.10
wrapt==1.16.0
wsproto==1.0.0
xformers==0.0.20
xkit==0.0.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
zope.interface==5.4.0
